Add AIController and AIManager classes.
An AIController has a pointer to the agent it controls. Calling its
run() method gives new steering values to the agent it controls.
An AIManager has a bunch of AIControllers. Its run method is what will
be called by the scheduler, and it exists so we know what agents should
be controlled by AI, as opposed to human input or network input.