Refactored Graphics::render() so that it uses the world singleton rather than taking a world argument
Also added a DrawCube method for debugging, and modified render(agent) to use it.
Finally, I renamed arrow to DrawArrow for consistency.

Currently, there is a bug in render(agent) wherein the velocity vector is being drawn correctly, but the agent
itself is drawn in the wrong place.  A cube and arrow are drawn where the agent should be (achieved by
avoiding matrix calculations) and then the sphere is drawn as before.