Convert a bunch of reference parameters to pointer ones.
I noticed that all of these functions (mostly for telling some kind of
manager class to look after the input) were taking a reference but just
converting it to a pointer to store it internally. And the reference
passing was causing memory initialization problems that the pointer
version don't seem to cause.